eth: move reward service into eth package
